+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - Fixes issue with xExchMaintenanceMode on Exchange 2016 where the cluster
   does not get paused when going into maintenance mode. Also fixes issue
   where services fail to stop, start, pause, or resume.
+- Explicitly cast member types in Get-DscConfiguration return hashtables to
+  align with the types defined in the resource schemas. Fixes an issue where
+  Get-DscConfiguration fails to return a value.
+- xExchClientAccessServer: Fixes issue where AlternateServiceAccountConfiguration
+  or RemoveAlternateServiceAccountCredentials parameters can't be used at the
+  same time as other optional parameters.
 
 ## 1.22.0.0